7.pizza_redis_pipeline/second-attempt/api/requirements.txt
fastapi
uvicorn
python-multipart
pymongo
redis
confluent-kafka
python-dotenv
mysql-connector-python
requests
